dev(narugo): fix existing one

+++ cyberharem/train/train.py
@@ -344,8 +344,11 @@ def train_lora(ds_repo_id: str, dataset_name: Optional[str] = None, workdir: Opt
         with load_reg_dataset(bangumi_repo_id=meta['bangumi'], bangumi_prefix_tag=bangumi_style_tag,
                               use_reg=use_reg, latent_cache_id=latent_cache_id) as reg_dir:
             features_path = os.path.join(workdir, 'features.npy')
-            logging.info(f'Extracting features from {train_dir!r}, and saving that to {features_path!r} ...')
-            np.save(features_path, _extract_features_from_directory(train_dir))
+            if os.path.exists(features_path):
+                logging.info('Features already exist, use the existing one.')
+            else:
+                logging.info(f'Extracting features from {train_dir!r}, and saving that to {features_path!r} ...')
+                np.save(features_path, _extract_features_from_directory(train_dir))
 
             r_boy, r_girl = _gender_predict(train_dir)
             if r_boy >= 0.7 or r_boy - r_girl >= 0.25:
